Refactor and consolidate integration tests for document and image handling in LLM prompts. Removed outdated tests and added new tests for document and image reference creation, token substitution, and extraction into user_documents and user_images lists. Enhanced structure for clarity and maintainability.
